value: remove CqlValueDisplayer
Implemented std::fmt::Display for CqlValue instead.
